Consolidating the test classes into the utilities.py script
